Move files in appropriate folders, remove old ones
Move all appropriate includes. Rename usercreation.ui
to page_usersetup.ui, replacing the old UI file.
