+++ src/modules/users/jobs/SetPasswordJob.cpp
@@ -16,7 +16,7 @@
  *   along with Calamares. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <SetPasswordJob.h>
+#include <jobs/SetPasswordJob.h>
 
 #include "JobQueue.h"
 #include "GlobalStorage.h"
